restructure chain_builder to make things cleaner
