+++ angrop/chain_builder/reg_mover.py
@@ -14,9 +14,9 @@ class RegMover(Builder):
     """
     handle register moves such as `mov rax, rcx`
     """
-    def __init__(self, project, arch, gadgets, badbytes=None, filler=None):
-        super().__init__(project, arch, badbytes=badbytes, filler=filler)
-        self._reg_moving_gadgets = self._filter_gadgets(gadgets)
+    def __init__(self, chain_builder):
+        super().__init__(chain_builder)
+        self._reg_moving_gadgets = self._filter_gadgets(self.chain_builder.gadgets)
 
     def verify(self, chain, preserve_regs, registers):
         """
@@ -69,7 +69,7 @@ class RegMover(Builder):
 
     def run(self, preserve_regs=None, **registers):
         if len(registers) == 0:
-            return RopChain(self.project, None, badbytes=self._badbytes)
+            return RopChain(self.project, None, badbytes=self.badbytes)
 
         # sanity check
         preserve_regs = set(preserve_regs) if preserve_regs else set()
